feat: graph render loading indicator, auto-center graph when rendered
