+++ packages/license-kit/visualizer/src/app/page.tsx
@@ -2,12 +2,12 @@
 
 import React from 'react';
 
+import DependencyGraph from '../components/DependencyGraph';
 import ReportPickerArea from '../components/ReportPickerArea';
-import Tree from '../components/Tree';
 import { useVisualizerStore } from '../store/visualizerStore';
 
 export default function App() {
   const { report } = useVisualizerStore();
 
-  return <ReportPickerArea>{report ? <Tree data={report} /> : null}</ReportPickerArea>;
+  return <ReportPickerArea>{report ? <DependencyGraph data={report} /> : null}</ReportPickerArea>;
 }
