+++ packages/license-kit/visualizer/src/constants.ts
@@ -1,3 +1,7 @@
+import type { Types } from '@callstack/licenses';
+
+import { buildPackageKey } from './utils/packageUtils';
+
 export const SIDEBAR_EXPANDED_WIDTH = 550;
 export const SIDEBAR_COLLAPSED_WIDTH = 50;
 
@@ -12,4 +16,12 @@ export const MAX_TREE_PARSING_DEPTH = 300;
 
 export const LABEL_FONT_SIZE = 14;
 
-export const ROOT_PACKAGE_KEY = 'root@0.0.0';
+export const ROOT_PROJECT_ROOT_PACKAGE: Types.License = {
+  name: 'root',
+  dependencyType: 'dependency',
+  version: '0.0.0',
+  requiredVersion: '0.0.0',
+  parentPackages: [],
+};
+
+export const ROOT_PROJECT_ROOT_PACKAGE_KEY = buildPackageKey(ROOT_PROJECT_ROOT_PACKAGE);
