+++ packages/license-kit/visualizer/src/providers/ClientSideProviders.tsx
@@ -1,13 +1,25 @@
 'use client';
 
-import { theme } from '@/styles/theme';
+import { useAppStore } from '@/store/appStore';
+import { darkTheme, lightTheme } from '@/styles/theme';
 import { CssBaseline, ThemeProvider } from '@mui/material';
 import { SnackbarProvider } from 'notistack';
-import React from 'react';
+import React, { useEffect, useMemo } from 'react';
 
 export type AppProvidersProps = React.PropsWithChildren;
 
 export function ClientSideProviders({ children }: AppProvidersProps) {
+  const { themeMode, setThemeMode } = useAppStore();
+
+  const theme = useMemo(() => {
+    return themeMode === 'dark' ? darkTheme : lightTheme;
+  }, [themeMode]);
+
+  // initial render theme mode sync
+  useEffect(() => {
+    setThemeMode(window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light');
+  }, [setThemeMode]);
+
   return (
     <ThemeProvider theme={theme}>
       <SnackbarProvider>
