+++ packages/license-kit/visualizer/src/utils/buildDependencyGraph.ts
@@ -1,6 +1,11 @@
 'use client';
 
-import { DEFAULT_RADIUS, MAX_TREE_PARSING_DEPTH, ROOT_PACKAGE_KEY } from '@/constants';
+import {
+  DEFAULT_RADIUS,
+  MAX_TREE_PARSING_DEPTH,
+  ROOT_PROJECT_ROOT_PACKAGE,
+  ROOT_PROJECT_ROOT_PACKAGE_KEY,
+} from '@/constants';
 import type { TreeNode } from '@/types/TreeNode';
 import type { Types } from '@callstack/licenses';
 import { graphlib } from '@dagrejs/dagre';
@@ -8,10 +13,16 @@ import _ from 'lodash';
 
 import { buildPackageKey } from './packageUtils';
 
-export function buildHierarchy(_data: Types.AggregatedLicensesMapping): {
+export type DependencyGraphResult = {
   graph: graphlib.Graph<TreeNode>;
   incomplete: boolean;
-} {
+  rootPackageKey: string;
+};
+
+export function buildDependencyGraph(
+  _data: Types.AggregatedLicensesMapping,
+  _selectedRoot: Types.License = ROOT_PROJECT_ROOT_PACKAGE,
+): DependencyGraphResult {
   const data = _.cloneDeep(_data);
 
   const graph = new graphlib.Graph<TreeNode>().setGraph({
@@ -20,16 +31,13 @@ export function buildHierarchy(_data: Types.AggregatedLicensesMapping): {
   });
 
   // 1. Build the tree structure
+  const rootPackageKey = buildPackageKey(_selectedRoot);
   const root: TreeNode = {
     children: [],
     meta: {
-      name: 'root',
-      dependencyType: 'dependency',
-      version: '0.0.0',
-      requiredVersion: '0.0.0',
+      ..._selectedRoot,
       parentPackageKeys: [],
-      parentPackages: [],
-      key: ROOT_PACKAGE_KEY,
+      key: rootPackageKey,
     },
     rank: 0,
   };
@@ -38,12 +46,15 @@ export function buildHierarchy(_data: Types.AggregatedLicensesMapping): {
 
   mapping[root.meta.key] = root;
 
-  graph.setNode(ROOT_PACKAGE_KEY, root);
+  graph.setNode(rootPackageKey, root);
 
   // 2. Recursively add children
   let depth = 0;
   let incomplete = false;
-  while (Object.keys(data).length > 0) {
+  let anyChangesAppliedThisIter = true;
+  while (Object.keys(data).length > 0 && anyChangesAppliedThisIter) {
+    anyChangesAppliedThisIter = false; // reset the flag
+
     for (const [packageKey, packageInfo] of Object.entries(data)) {
       const parentPackageKeys = packageInfo.parentPackages.length
         ? packageInfo.parentPackages.map((parentPackageInfo) =>
@@ -52,7 +63,9 @@ export function buildHierarchy(_data: Types.AggregatedLicensesMapping): {
               version: parentPackageInfo.resolvedVersion,
             }),
           )
-        : [ROOT_PACKAGE_KEY];
+        : _selectedRoot === ROOT_PROJECT_ROOT_PACKAGE
+        ? [rootPackageKey]
+        : [];
 
       for (const parentPackageKey of parentPackageKeys) {
         if (parentPackageKey in mapping) {
@@ -77,12 +90,15 @@ export function buildHierarchy(_data: Types.AggregatedLicensesMapping): {
           graph.setEdge(parentPackageKey, packageKey, { label: packageInfo.dependencyType });
 
           delete data[packageKey]; // remove processed package
+
+          anyChangesAppliedThisIter = true;
         }
       }
     }
 
-    // failsafe
-    if (depth > MAX_TREE_PARSING_DEPTH) {
+    // failsafe for detecting non-full graph processed - should never happen; it's only possible to validate
+    // this in such a simple way as below if the selected root is the root itself (in which case all the packages should be included in the graph)
+    if (depth > MAX_TREE_PARSING_DEPTH && root.meta.key === ROOT_PROJECT_ROOT_PACKAGE_KEY) {
       console.warn(
         `Reached maximum tree parsing depth of ${MAX_TREE_PARSING_DEPTH}. Stopping to prevent infinite loop.`,
       );
@@ -94,5 +110,5 @@ export function buildHierarchy(_data: Types.AggregatedLicensesMapping): {
   }
 
   // 3. Create a d3 hierarchy
-  return { graph, incomplete };
+  return { graph, incomplete, rootPackageKey };
 }
