+++ packages/licenses-api/src/permissiveness/constants.ts
@@ -2,7 +2,7 @@ import { LicenseCategory } from './LicenseCategory';
 
 export const PERMISSIVENESS_SCORE_WEIGHTS = {
   [LicenseCategory.PERMISSIVE]: 50,
-  [LicenseCategory.WEAK_COPYLEFT]: 40,
-  [LicenseCategory.STRONG_COPYLEFT]: 30,
-  [LicenseCategory.UNKNOWN]: 10,
+  [LicenseCategory.WEAK_COPYLEFT]: 20,
+  [LicenseCategory.STRONG_COPYLEFT]: 0,
+  [LicenseCategory.UNKNOWN]: 30,
 };
